+++ libs/datapipe-core/datapipe/meta/base.py
@@ -5,11 +5,11 @@ import pandas as pd
 from sqlalchemy import Column
 
 from datapipe.run_config import RunConfig
-from datapipe.types import ChangeList, DataSchema, FieldAccessor, HashDF, IndexDF, MetadataDF, MetaSchema
+from datapipe.types import ChangeList, DataField, DataSchema, FieldAccessor, HashDF, IndexDF, MetadataDF, MetaSchema
 
 if TYPE_CHECKING:
-    from datapipe.compute import ComputeInput
-    from datapipe.datatable import DataStore, DataTable
+    from datapipe.compute import ComputeInput, ComputeOutput
+    from datapipe.datatable import DataStore
 
 
 class MetaPlane:
@@ -37,7 +37,7 @@ class MetaPlane:
         self,
         name: str,
         input_dts: Sequence["ComputeInput"],
-        output_dts: Sequence["DataTable"],
+        output_dts: Sequence["ComputeOutput"],
         transform_keys: list[str] | None = None,
         order_by: list[str] | None = None,
         order: Literal["asc", "desc"] = "asc",
@@ -178,7 +178,7 @@ class TableMeta:
         Given an index dataframe with transform keys, return an index dataframe
         with table keys, applying `keys` aliasing if provided.
 
-        * `keys` is a mapping from table key to transform key
+        * `keys` is a mapping from transform key to table key
         """
 
         if keys is None:
@@ -186,11 +186,11 @@ class TableMeta:
 
         table_key_cols: dict[str, pd.Series] = {}
         for transform_col in transform_idx.columns:
-            accessor = keys.get(transform_col) if keys is not None else transform_col
-            if isinstance(accessor, str):
-                table_key_cols[accessor] = transform_idx[transform_col]
-            else:
-                pass  # skip non-meta fields
+            table_col = keys.get(transform_col)
+            if isinstance(table_col, DataField):
+                continue
+            if table_col is not None:
+                table_key_cols[table_col] = transform_idx[transform_col]
 
         return IndexDF(pd.DataFrame(table_key_cols))
 
@@ -202,8 +202,8 @@ class TransformMeta:
     @classmethod
     def compute_transform_schema(
         cls,
-        input_cis: Sequence["ComputeInput"],
-        output_dts: Sequence["DataTable"],
+        inputs: Sequence["ComputeInput"],
+        outputs: Sequence["ComputeOutput"],
         transform_keys: list[str] | None,
     ) -> tuple[list[str], MetaSchema]:
         # Hacky way to collect all the primary keys into a single set. Possible
@@ -211,24 +211,24 @@ class TransformMeta:
         # same key is defined differently in different input tables.
         all_keys: dict[str, Column] = {}
 
-        for ci in input_cis:
+        for ci in inputs:
             all_keys.update({col.name: col for col in ci.primary_schema})
 
-        for dt in output_dts:
-            all_keys.update({col.name: col for col in dt.primary_schema})
+        for co in outputs:
+            all_keys.update({col.name: col for col in co.primary_schema})
 
         if transform_keys is not None:
             return (transform_keys, [all_keys[k] for k in transform_keys])
 
-        assert len(input_cis) > 0, "At least one input table is required to infer transform keys"
+        assert len(inputs) > 0, "At least one input table is required to infer transform keys"
 
-        inp_p_keys = set.intersection(*[set(inp.primary_keys) for inp in input_cis])
+        inp_p_keys = set.intersection(*[set(inp.primary_keys) for inp in inputs])
         assert len(inp_p_keys) > 0
 
-        if len(output_dts) == 0:
+        if len(outputs) == 0:
             return (list(inp_p_keys), [all_keys[k] for k in inp_p_keys])
 
-        out_p_keys = set.intersection(*[set(out.primary_keys) for out in output_dts])
+        out_p_keys = set.intersection(*[set(out.primary_keys) for out in outputs])
         assert len(out_p_keys) > 0
 
         inp_out_p_keys = set.intersection(inp_p_keys, out_p_keys)