+++ libs/datapipe-core/datapipe/meta/sql_meta.py
@@ -16,8 +16,7 @@ import pandas as pd
 import sqlalchemy as sa
 from opentelemetry import trace
 
-from datapipe.compute import ComputeInput
-from datapipe.datatable import DataTable
+from datapipe.compute import ComputeInput, ComputeOutput
 from datapipe.meta.base import MetaPlane, TableDebugInfo, TableMeta, TransformMeta
 from datapipe.run_config import LabelDict, RunConfig
 from datapipe.sql_util import sql_apply_idx_filter_to_table, sql_apply_runconfig_filter
@@ -26,9 +25,7 @@ from datapipe.store.table_store import TableStore
 from datapipe.types import (
     ChangeList,
     DataDF,
-    DataField,
     DataSchema,
-    FieldAccessor,
     HashDF,
     IndexDF,
     MetadataDF,
@@ -357,7 +354,7 @@ class SQLTableMeta(TableMeta):
         self,
         transform_keys: list[str],
         table_store: TableStore,
-        keys: dict[str, FieldAccessor],
+        keys: dict[str, str],
         filters_idx: IndexDF | None = None,
         run_config: RunConfig | None = None,
     ) -> tuple[list[str], Any]:
@@ -365,9 +362,7 @@ class SQLTableMeta(TableMeta):
         Create a CTE that aggregates the table by transform keys, applies keys
         aliasing and returns the maximum update_ts for each group.
 
-        * `keys` is a mapping from transform key to table key accessor
-          (can be string for meta table column or DataField for data table
-          column)
+        * `keys` is a mapping from transform key to meta table column name
         * `transform_keys` is a list of keys used in the transformation
 
         CTE has the following columns:
@@ -379,40 +374,20 @@ class SQLTableMeta(TableMeta):
         present in primary keys of this CTE
         """
 
-        from datapipe.store.database import TableStoreDB
-
         meta_table = self.sql_table
-        data_table = None
 
         key_cols: list[Any] = []
         cte_transform_keys: list[str] = []
-        should_join_data_table = False
 
         for transform_key in transform_keys:
-            # TODO convert to match when we deprecate Python 3.9
             accessor = keys.get(transform_key, transform_key)
-            if isinstance(accessor, str):
-                if accessor in self.primary_keys:
-                    key_cols.append(meta_table.c[accessor].label(transform_key))
-                    cte_transform_keys.append(transform_key)
-            elif isinstance(accessor, DataField):
-                should_join_data_table = True
-                assert isinstance(table_store, TableStoreDB)
-                data_table = table_store.data_table
-
-                key_cols.append(data_table.c[accessor.field_name].label(transform_key))
+            if accessor in self.primary_keys:
+                key_cols.append(meta_table.c[accessor].label(transform_key))
                 cte_transform_keys.append(transform_key)
 
         sql: Any = sa.select(*key_cols + [sa.func.max(meta_table.c["update_ts"]).label("update_ts")]).select_from(
             meta_table
         )
-
-        if should_join_data_table:
-            assert data_table is not None
-            sql = sql.join(
-                data_table,
-                sa.and_(*[meta_table.c[pk] == data_table.c[pk] for pk in self.primary_keys]),
-            )
 
         if len(key_cols) > 0:
             sql = sql.group_by(*key_cols)
@@ -436,8 +411,8 @@ class SQLTransformMeta(TransformMeta):
         self,
         dbconn: DBConn,
         name: str,
-        input_cis: Sequence[ComputeInput],
-        output_dts: Sequence[DataTable],
+        inputs: Sequence[ComputeInput],
+        outputs: Sequence[ComputeOutput],
         transform_keys: list[str] | None,
         order_by: list[str] | None = None,
         order: Literal["asc", "desc"] = "asc",
@@ -446,12 +421,12 @@ class SQLTransformMeta(TransformMeta):
         self.dbconn = dbconn
         self.name = name
 
-        self.input_cis = input_cis
-        self.output_dts = output_dts
+        self.inputs = inputs
+        self.outputs = outputs
 
         self.transform_keys, self.transform_keys_schema = self.compute_transform_schema(
-            input_cis=self.input_cis,
-            output_dts=self.output_dts,
+            inputs=self.inputs,
+            outputs=self.outputs,
             transform_keys=transform_keys,
         )
 
@@ -476,8 +451,8 @@ class SQLTransformMeta(TransformMeta):
         return self.__class__, (
             self.dbconn,
             self.name,
-            self.input_cis,
-            self.output_dts,
+            self.inputs,
+            self.outputs,
             self.transform_keys,
             self.order_by,
             self.order,
@@ -504,7 +479,7 @@ class SQLTransformMeta(TransformMeta):
         run_config: RunConfig | None = None,
     ) -> tuple[int, Iterable[IndexDF]]:
         with tracer.start_as_current_span("compute ids to process"):
-            if len(self.input_cis) == 0:
+            if len(self.inputs) == 0:
                 return (0, iter([]))
 
             idx_count = self.get_changed_idx_count(
@@ -549,7 +524,7 @@ class SQLTransformMeta(TransformMeta):
         with tracer.start_as_current_span("compute ids to process"):
             changes = [pd.DataFrame(columns=self.transform_keys)]
 
-            for inp in self.input_cis:
+            for inp in self.inputs:
                 if inp.dt.name in change_list.changes:
                     idx = change_list.changes[inp.dt.name]
                     if any([key not in idx.columns for key in self.transform_keys]):
@@ -747,11 +722,11 @@ class SQLTransformMeta(TransformMeta):
         run_config: RunConfig | None = None,  # TODO remove
     ) -> Any:
         all_input_keys_counts: dict[str, int] = {}
-        for col in itertools.chain(*[inp.dt.primary_schema for inp in self.input_cis]):
+        for col in itertools.chain(*[inp.dt.primary_schema for inp in self.inputs]):
             all_input_keys_counts[col.name] = all_input_keys_counts.get(col.name, 0) + 1
 
         inp_ctes = []
-        for inp in self.input_cis:
+        for inp in self.inputs:
             inp_meta = inp.dt.meta
             assert isinstance(inp_meta, SQLTableMeta)
 
@@ -902,16 +877,16 @@ def _make_agg_of_agg(
                     onclause.append(prev_cte.cte.c[key] == cte.cte.c[key])
 
         if len(onclause) > 0:
-            sql = sql.outerjoin(
+            sql = sql.join(
                 cte.cte,
                 onclause=sa.and_(*onclause),
-                full=True,
+                full=cte.join_type == "full",
             )
         else:
-            sql = sql.outerjoin(
+            sql = sql.join(
                 cte.cte,
                 onclause=sa.literal(True),
-                full=True,
+                full=cte.join_type == "full",
             )
 
         if cte.join_type == "inner":
@@ -949,7 +924,7 @@ class SQLMetaPlane(MetaPlane):
         self,
         name: str,
         input_dts: Sequence[ComputeInput],
-        output_dts: Sequence[DataTable],
+        output_dts: Sequence[ComputeOutput],
         transform_keys: list[str] | None = None,
         order_by: list[str] | None = None,
         order: Literal["asc", "desc"] = "asc",
@@ -957,8 +932,8 @@ class SQLMetaPlane(MetaPlane):
         return SQLTransformMeta(
             dbconn=self.dbconn,
             name=name,
-            input_cis=input_dts,
-            output_dts=output_dts,
+            inputs=input_dts,
+            outputs=output_dts,
             transform_keys=transform_keys,
             order_by=order_by,
             order=order,